refactor(Repeater): Split ItemsRepeaterScrollHost per winui-port
Splits ItemsRepeaterScrollHost into the standard .cs / .h.mux.cs / .mux.cs trio. Body
follows WinUI commit 4b206bce3 and keeps the existing Uno-specific
VisualTreeHelper-based child slot (Uno's ItemsRepeaterScrollHost inherits from
FrameworkElement rather than Panel). The active anchoring code path matches the WinUI
non-RS5 fallback path 1:1.

Tracks issue #22908.
